securing routes

+++ src/Controller/AdminController.php
@@ -21,56 +21,61 @@ class AdminController extends AbstractController
     #[Route('/admin/deleteSpot/{id}', name: 'deleteSpot_admin')]
     public function deleteSpot(Security $security, EntityManagerInterface $entityManager, ManagerRegistry $doctrine, Spot $spot = null, Request $request): Response
     {
-        // Définition du user
+        // définition du $user actif
         $user=$this->getUser();
-            // Si un user est connecté && qu'il a le role admin
-            if($user && $user->getRoles() === '["ROLE_ADMIN"]'){
+        $userRole = $user->getRole();
+        // Si le $user possède le role admin
+        if ($user && in_array('ROLE_ADMIN', $userRole)) {
                 if ($spot){
                     $entityManager->remove($spot);
                     $entityManager->flush();
     
                 }
-            }else{
-                throw new \Exception('Accès reservé aux administrateurs connectés.');
-                return $this->redirectToRoute('app_home');
-            }
-            return $this->redirectToRoute('app_admin');
+                return $this->redirectToRoute('app_admin');
+        }else{
+            return $this->redirectToRoute('app_home');
+        }
             
     }
     
     // Pour supprimer une photo d'un spot
     #[Route('/admin/deletePicture/{idSpot}/{idPic}', name: 'deletePicture_admin')]
-    public function deletePicture(Security $security, Spot $idSpot, Picture $idPic , ManagerRegistry $doctrine) : Response
+    public function deletePicture(Spot $idSpot, Picture $idPic , ManagerRegistry $doctrine) : Response
     {   
-        $spot = $idSpot;
-        $picture = $idPic;
+        // définition du $user actif
+        $userActive=$this->getUser();
+        $userRole = $userActive->getRoles();
+        // Si le $user possède le role admin
+        if ($userActive && in_array('ROLE_ADMIN', $userRole)) {
+            // Définition des variables
+            $spot = $idSpot;
+            $picture = $idPic;
+            $user = $this->getUser();
+            // Définition du rôle 
+            $userRole = $user->getRoles();
+            // Définition du manager de doctrine
+            $entityManager = $doctrine->getManager();
 
-        $user = $security->getUser();
-        // Définition du rôle 
-        $userRole = $user->getRoles();
-        // Définition du manager de doctrine
-        $entityManager = $doctrine->getManager();
-        // dd($userRole);
-        //si le user est admin
-        if(in_array('ROLE_ADMIN', $userRole)){
             // Supprimer l'entrée de Picture correspondante
             $entityManager->remove($picture);
             $entityManager->flush();
-            
-        }
 
-       return $this->redirectToRoute('show_spot', ["idSpot"=>$spot->getId()]);
+            return $this->redirectToRoute('show_spot', ["idSpot"=>$spot->getId()]);
+        }else{
+            return $this->redirectToRoute('app_home');
+        }
     }
 
     #[Route('/admin/validate/{id}', name: 'validateSpot_admin')]
     public function validateSpot(Security $security, EntityManagerInterface $entityManager, Spot $spot)
     {
-        //RESTE A GERER LES CONDITIONS ADMINS
-        // $user=$security->getUser();
-
-        //Selectionne le spot a modifier en pointant l'id
-        $myspot = $entityManager->getRepository(Spot::class)->findOneBy(['id'=>$spot]);
-
+        // définition du $user actif
+        $userActive=$this->getUser();
+        $userRole = $userActive->getRoles();
+        // Si le $user possède le role admin
+        if ($userActive && in_array('ROLE_ADMIN', $userRole)) {
+            //Selectionne le spot a modifier en pointant l'id
+            $myspot = $entityManager->getRepository(Spot::class)->findOneBy(['id'=>$spot]);
             // Si le isValidated == FALSE -> le passe en TRUE
             if ($myspot->getIsValidated() == false) {
                 // Modifier la propriété de l'entité
@@ -89,19 +94,28 @@ class AdminController extends AbstractController
                 // Persist les modifications dans la base de données
                 $entityManager->flush();
             }
-        return $this->redirectToRoute('app_admin');
+            return $this->redirectToRoute('app_admin');
+        }else{
+            return $this->redirectToRoute('app_home');
+        }
     }
     #[Route('/admin/modify/{id}', name: 'modifySpot_admin')]
     public function modifySpot(Security $security, EntityManagerInterface $entityManager, Spot $spot)
     {
-        //RESTE A GERER LES CONDITIONS ADMINS
-        // $user=$security->getUser();
+        // définition du $user actif
+        $userActive=$this->getUser();
+        $userRole = $userActive->getRoles();
+        // Si le $user possède le role admin
+        if ($userActive && in_array('ROLE_ADMIN', $userRole)) {
+            //Selectionne le spot a modifier en pointant l'id
+            $myspot = $entityManager->getRepository(Spot::class)->findOneBy(['id'=>$spot]);
 
-        //Selectionne le spot a modifier en pointant l'id
-        $myspot = $entityManager->getRepository(Spot::class)->findOneBy(['id'=>$spot]);
-
-            
-        return $this->redirectToRoute('app_admin');
+                
+            return $this->redirectToRoute('app_admin');
+        }
+        else{
+            return $this->redirectToRoute('app_home');
+        }
     }
     
 
@@ -109,19 +123,29 @@ class AdminController extends AbstractController
     #[Route('/admin/listUsers', name: 'listUsers_admin')]
     public function listUsers(Security $security, ManagerRegistry $doctrine, Request $request)
     {
-        $users = $doctrine->getRepository(User::class)->findBy([], ['registrationDate'=>'DESC']);
+        // définition du $user actif
+        $userActive=$this->getUser();
+        $userRole = $userActive->getRoles();
+        // Si le $user possède le role admin
+        if ($userActive && in_array('ROLE_ADMIN', $userRole)) {
+            $listUsers = $doctrine->getRepository(User::class)->findBy([], ['registrationDate'=>'DESC']);
 
-        return $this->render('admin/listUsers.html.twig', [
-            'controller_name' => 'AdminController',
-            'users' => $users
-        ]);
+            return $this->render('admin/listUsers.html.twig', [
+                'controller_name' => 'AdminController',
+                'users' => $listUsers
+            ]);
+        }else{
+            return $this->redirectToRoute('app_home');
+        }
     }
     #[Route('/admin/banUser/{id}', name: 'banUser_admin')]
     public function banUser(Security $security, EntityManagerInterface $entityManager, ManagerRegistry $doctrine, User $user = null, Request $request): Response
     {
-        //RESTE A GERER LES CONDITIONS ADMINS
-        // $user=$security->getUser();
-
+        // définition du $user actif
+        $userActive=$this->getUser();
+        $userRole = $userActive->getRoles();
+        // Si le $user possède le role admin
+        if ($userActive && in_array('ROLE_ADMIN', $userRole)) {
             if ($user && $user->isIsBanned() == 0){
                 $user->setIsBanned(1);
             }else{
@@ -129,43 +153,60 @@ class AdminController extends AbstractController
             }
             $entityManager->flush();
                 return $this->redirectToRoute('listUsers_admin');
+        }else{
+            return $this->redirectToRoute('app_home');
+        }
     }
     #[Route('/admin/deleteUser/{id}', name: 'deleteUser_admin')]
     public function deleteUser(Security $security, EntityManagerInterface $entityManager, ManagerRegistry $doctrine, User $user = null, Request $request): Response
     {
-    //RESTE A GERER LES CONDITIONS ADMINS
-    // $user=$security->getUser();
-        if ($user){
+        // définition du $user actif
+        $userActive=$this->getUser();
+        $userRole = $userActive->getRoles();
+        // Si le $user possède le role admin
+        if ($userActive && in_array('ROLE_ADMIN', $userRole)) {
             $entityManager->remove($user);
             $entityManager->flush();
+            return $this->redirectToRoute('listUsers_admin');
+        }else{
+            return $this->redirectToRoute('app_home');
         }
-        return $this->redirectToRoute('listUsers_admin');
     }
 //***********************************CRUD MODULE****************** */
-#[Route('/admin/deleteModule/{id}', name: 'deleteModule_admin')]
-public function deleteModule(Security $security, EntityManagerInterface $entityManager, ManagerRegistry $doctrine, Module $module = null, Request $request): Response
-{
-    //RESTE A GERER LES CONDITIONS ADMINS
-    // $user=$security->getUser();
-
-    
-        if ($module){
-            $entityManager->remove($module);
-            $entityManager->flush();
+    #[Route('/admin/deleteModule/{id}', name: 'deleteModule_admin')]
+    public function deleteModule(Security $security, EntityManagerInterface $entityManager, ManagerRegistry $doctrine, Module $module = null, Request $request): Response
+    {
+        // définition du $user actif
+        $user=$this->getUser();
+        $userRole = $user->getRoles();
+        // Si le $user possède le role admin
+        if ($user && in_array('ROLE_ADMIN', $userRole)) {
+            if ($module){
+                $entityManager->remove($module);
+                $entityManager->flush();
+            }
+            return $this->redirectToRoute('app_module');
+        }else{
+            return $this->redirectToRoute('app_home');
         }
-        return $this->redirectToRoute('app_module');
     }
 
 
     #[Route('/admin', name: 'app_admin')]
     public function index(ManagerRegistry $doctrine, Spot $spot = null, Request $request): Response
     {
-        $spots = $doctrine->getRepository(Spot::class)->findBy([], ['creationDate'=>'DESC']);
-
-        return $this->render('admin/index.html.twig', [
-            'controller_name' => 'AdminController',
-            'spots' => $spots
-        ]);
+        // définition du $user actif
+        $user=$this->getUser();
+        $userRole = $user->getRoles();
+        // Si le $user possède le role admin
+        if ($user && in_array('ROLE_ADMIN', $userRole)) {
+            $spots = $doctrine->getRepository(Spot::class)->findBy([], ['creationDate'=>'DESC']);
+            return $this->render('admin/index.html.twig', [
+                'spots' => $spots
+            ]);
+        }else{
+            return $this->redirectToRoute('app_home');
+        }
     }
 }
